+++ CHANGES.rst
@@ -5,16 +5,20 @@ TestLink-API-Python-client v0.6.2 - Under Develop
 ------------------------------------------------------------
 support for TL 1.9.14 release
 
-implement 1.9.14 new api interfaces - #53
+implement 1.9.14 new api interfaces - #53 #61
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 new TestlinkAPIGeneric and TestlinkAPIClient api methods
 
 - deleteTestProject(<testprojectprefix>) 
+- updateTestSuiteCustomFieldDesignValue(<testprojectid>, <testsuiteid>, <customfields>)
 
 examples _`<example/TestLinkExample.py>`_  and 
 `<example/TestLinkExampleGenericApi.py>`_ now deletes the sample project, if it
 already exist.
+
+example _`<example/TestLinkExample_CF_KW.py>`_ shows, how to set and get 
+customer field values
 
 implement 1.9.14 changed api interfaces - #48 #49 #54 #59
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -74,6 +78,11 @@ configuration
 TestlinkAPIClient service method listKeywordsForTC() uses now getTestCaseKeywords()
 
 - internal change to reduce code complexity 
+
+Known TL 1.9.14 limitations:
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+- 7282 updateTestSuiteCustomFieldDesignValue() does change customer field values
 
 TestLink-API-Python-client release notes v0.6.1 (Mar. 2015)  
 ------------------------------------------------------------